refac: move popup to views
